Move `deriveKeys` into `src/utils/internal` folder.
