+++ src/utils/decryptAES.ts
@@ -15,7 +15,7 @@
  */
 
 import { decodeBase64 } from "../base64.ts";
-import { deriveKeys } from "../rust-crypto/deriveKeys.ts";
+import { deriveKeys } from "./internal/deriveKeys.ts";
 import { AESEncryptedSecretStoragePayload } from "../@types/AESEncryptedSecretStoragePayload.ts";
 
 /**
